+++ temp-seed-scripts/index.js
@@ -1,19 +1,21 @@
+// ! EVERYTHING COMMENTED OUT EXCEPT LIVES HISTORY SEED
+
 const connectToMongoDb = require('../connect-to-mongodb')
 const _ = require('lodash')
 
-const cleanQualityAccessPtps = require('./cleanQualityAccessPtps')
+// const cleanQualityAccessPtps = require('./cleanQualityAccessPtps')
 
-const seedLines = require('./seedLines')
-const seedPopulations = require('./seedPopulation')
+// const seedLines = require('./seedLines')
+// const seedPopulations = require('./seedPopulation')
 // const seedNewIndications = require('./seedNewIndications')
 // const seedNewRegimens = require('./seedNewRegimens')
 
-const seedTreatmentPlans = require('./seedTreatmentPlans')
+// const seedTreatmentPlans = require('./seedTreatmentPlans')
 
-const seedOrganizationsTreatmentPlans = require('./seedOrganizationsTreatmentPlans')
-const seedOrganizationsTreatmentPlansHistory = require('./seedOrganizationsTreatmentPlansHistory')
+// const seedOrganizationsTreatmentPlans = require('./seedOrganizationsTreatmentPlans')
+// const seedOrganizationsTreatmentPlansHistory = require('./seedOrganizationsTreatmentPlansHistory')
 
-const seedTdgProjectsCollection = require('./seedTdgProjectsCollection')
+// const seedTdgProjectsCollection = require('./seedTdgProjectsCollection')
 
 const seedLivesCollection = require('./seedLivesCollection')
 
@@ -23,17 +25,17 @@ const runSeedScripts = async () => {
 
   const pulseCore = dbs.db('pulse-core')
 
-  await cleanQualityAccessPtps(pulseCore)
+  // await cleanQualityAccessPtps(pulseCore)
 
   const [
-    payerHistoricalQualityAccess,
-    payerHistoricalAdditionalCriteria,
-    payerHistoricalPolicyLinks,
+    // payerHistoricalQualityAccess,
+    // payerHistoricalAdditionalCriteria,
+    // payerHistoricalPolicyLinks,
     payerOrganizations,
   ] = await Promise.all([
-    pulseCore.collection('payerHistoricalQualityAccess').find({}).toArray(),
-    pulseCore.collection('payerHistoricalAdditionalCriteria').find({}).toArray(),
-    pulseCore.collection('payerHistoricalPolicyLinks').find({}).toArray(),
+    // pulseCore.collection('payerHistoricalQualityAccess').find({}).toArray(),
+    // pulseCore.collection('payerHistoricalAdditionalCriteria').find({}).toArray(),
+    // pulseCore.collection('payerHistoricalPolicyLinks').find({}).toArray(),
     pulseCore.collection('organizations').find({ type: 'Payer' }).toArray(),
   ])
 
@@ -41,30 +43,30 @@ const runSeedScripts = async () => {
 
   const seedParameters = {
     pulseCore,
-    payerHistoricalQualityAccess,
-    payerHistoricalAdditionalCriteria,
-    payerHistoricalPolicyLinks,
+    // payerHistoricalQualityAccess,
+    // payerHistoricalAdditionalCriteria,
+    // payerHistoricalPolicyLinks,
     payerOrganizationsBySlug,
   }
 
   console.log(`Historical docs loaded\nBeginning seeding\n`);
 
   // 1. seed larger sets of treatment plan parts (ones we can't just hard-code like coverages or books)
-  await Promise.all([
-    seedLines(seedParameters),
-    seedPopulations(seedParameters),
-  ])
+  // await Promise.all([
+  //   seedLines(seedParameters),
+  //   seedPopulations(seedParameters),
+  // ])
 
-  // 2. seed actual treatment plan combos
-  await seedTreatmentPlans(seedParameters)
+  // // 2. seed actual treatment plan combos
+  // await seedTreatmentPlans(seedParameters)
 
-  await seedOrganizationsTreatmentPlans(seedParameters)
+  // await seedOrganizationsTreatmentPlans(seedParameters)
 
   await seedLivesCollection(seedParameters)
 
-  await seedOrganizationsTreatmentPlansHistory(seedParameters)
+  // await seedOrganizationsTreatmentPlansHistory(seedParameters)
 
-  await seedTdgProjectsCollection(seedParameters)
+  // await seedTdgProjectsCollection(seedParameters)
 
   dbs.close()
 }