+++ temp-seed-scripts/seedLivesCollection/index.js
@@ -1,5 +1,3 @@
-const _ = require('lodash')
-
 const getHistoricalDocs = require('./getHistoricalDocs')
 
 // ! MSA lives are NOT handled by this script and are being ignored for now