+++ temp-seed-scripts/seedLivesCollection/key-book-coverage-map.js
@@ -1,50 +1,52 @@
+const { ObjectId } = require('mongodb')
+
 module.exports = {
   commercialMedical: {
-    book: 'Commercial',
-    coverage: 'Medical',
+    bookId: ObjectId("5ea6ef76e4e250454c670a3e"),
+    coverageId: ObjectId("5ebdafb51594eb3c26f4d562"),
   },
   commercialPharmacy: {
-    book: 'Commercial',
-    coverage: 'Pharmacy',
+    bookId: ObjectId("5ea6ef76e4e250454c670a3e"),
+    coverageId: ObjectId("5ebdafc81594eb3c26f4d563"),
   },
   ffsMedicaidMedical: {
-    book: 'FFS Medicaid',
-    coverage: 'Medical',
+    bookId: ObjectId("5ea6ef76e4e250454c670a3f"),
+    coverageId: ObjectId("5ebdafb51594eb3c26f4d562"),
   },
   ffsMedicaidPharmacy: {
-    book: 'FFS Medicaid',
-    coverage: 'Pharmacy',
+    bookId: ObjectId("5ea6ef76e4e250454c670a3f"),
+    coverageId: ObjectId("5ebdafc81594eb3c26f4d563"),
   },
   macMedical: {
-    book: 'MAC',
-    coverage: 'Medical',
+    bookId: ObjectId("5ea6ef76e4e250454c670a43"),
+    coverageId: ObjectId("5ebdafb51594eb3c26f4d562"),
   },
   managedMedicaidMedical: {
-    book: 'Managed Medicaid',
-    coverage: 'Medical',
+    bookId: ObjectId("5ea6ef76e4e250454c670a40"),
+    coverageId: ObjectId("5ebdafb51594eb3c26f4d562"),
   },
   managedMedicaidPharmacy: {
-    book: 'Managed Medicaid',
-    coverage: 'Pharmacy',
+    bookId: ObjectId("5ea6ef76e4e250454c670a40"),
+    coverageId: ObjectId("5ebdafc81594eb3c26f4d563"),
   },
   medicareMedical: {
-    book: 'Medicare',
-    coverage: 'Medical',
+    bookId: ObjectId("5ea6ef76e4e250454c670a41"),
+    coverageId: ObjectId("5ebdafb51594eb3c26f4d562"),
   },
   medicarePharmacy: {
-    book: 'Medicare',
-    coverage: 'Pharmacy',
+    bookId: ObjectId("5ea6ef76e4e250454c670a41"),
+    coverageId: ObjectId("5ebdafc81594eb3c26f4d563"),
   },
   federalOtherMedical: {
-    book: 'Federal/Other',
-    coverage: 'Medical',
+    bookId: ObjectId("5ea6ef76e4e250454c670a42"),
+    coverageId: ObjectId("5ebdafb51594eb3c26f4d562"),
   },
   federalOtherPharmacy: {
-    book: 'Federal/Other',
-    coverage: 'Pharmacy',
+    bookId: ObjectId("5ea6ef76e4e250454c670a42"),
+    coverageId: ObjectId("5ebdafc81594eb3c26f4d563"),
   },
   healthExchangePharmacy: {
-    book: 'Health Exchange',
-    coverage: 'Pharmacy',
+    bookId: ObjectId("5ea6ef76e4e250454c670a45"),
+    coverageId: ObjectId("5ebdafc81594eb3c26f4d563"),
   },
 }