+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - Started: 7-17-20
 
+### Changed
+- `Polaris Scripts`: Seeding of `lives.history` now persists bookId and coverageId, instead of just string references.
+
 ## [3.0.0] - Started: 7-9-20
 
 ### Added